+++ CHANGELOG.md
@@ -10,7 +10,10 @@ changelog does not include internal changes that do not affect the user.
 
 ### Added
 
-- Added `GradVac` and `GradVacWeighting` from [Gradient Vaccine: Investigating and Improving Multi-task Optimization in Massively Multilingual Models](https://arxiv.org/pdf/2010.05874).
+- Added `GradVac` and `GradVacWeighting` from 
+  [Gradient Vaccine: Investigating and Improving Multi-task Optimization in Massively Multilingual Models](https://arxiv.org/pdf/2010.05874).
+- Added a fallback for when the inner optimization of `NashMTL` fails (which can happen for example
+  on the matrix [[0., 0.], [0., 1.]]).
 
 ## [0.9.0] - 2026-02-24